Create Check Windows Hello for Business status using registry
This PowerShell script checks the status of Windows Hello for Business on a device by looking at the registry key HKLM:\SOFTWARE\Policies\Microsoft\PassportForWork. If the key is present and the value of the Enabled property is set to 1, then Windows Hello for Business is enabled. Otherwise, it is not enabled.
This script can be useful for administrators who need to check the status of Windows Hello for Business on multiple devices.
To use the script, you will need to have PowerShell installed on the device and you will need to run the script as an administrator.